pyproject.toml
[tool.poetry]
name = "ovpn-portal"
version = "0.0.0"
description = "A secure OpenVPN authentication portal with Google OAuth2"
authors = ["Abigail Ranson <abby@abbyranson.com>"]
readme = "README.md"
homepage = "https://github.com/ranson21/ovpn-client-web"
repository = "https://github.com/ranson21/ovpn-client-web"
documentation = "https://github.com/ranson21/ovpn-client-web#readme"
keywords = ["vpn", "openvpn", "oauth", "google-auth", "flask"]
license = "MIT"
packages = [{ include = "ovpn_portal" }]
classifiers = [
  "Development Status :: 4 - Beta",
  "Environment :: Web Environment",
  "Framework :: Flask",
  "Intended Audience :: System Administrators",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Topic :: System :: Systems Administration",
  "Topic :: Security",
]

[tool.poetry.dependencies]
python = "^3.8"
Flask = "^2.0.0"
flask-cors = "^4.0.0"
google-auth = "^2.0.0"
requests = "^2.31.0"
python-dotenv = "^0.19.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0.0"
pytest-cov = "^4.1.0"
black = "^22.0.0"
flake8 = "^4.0.0"
isort = "^5.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
ovpn-portal = "ovpn_portal.run:main"
